requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python-headless
pillow
numpy
python-multipart
huggingface_hub>=0.23
requests>=2.31